Implement flag wrapper to mockserver and its response
